W3_MCP/mcp-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finint-mcp-server"
version = "1.0.0"
description = "Indian Financial Intelligence MCP Server"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.6.0",
    "crewai>=1.12.2,<1.13",
    "crewai-tools>=0.35.0",
    "openai>=1.60.0",
    "langchain-openai>=0.3.0",
    "langsmith>=0.3.0",
    "pyjwt[crypto]>=2.9.0",
    "cryptography>=44.0.0",
    "httpx>=0.28.0",
    "redis>=5.2.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "structlog>=24.4.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "cachetools>=5.5.0",
    "smartapi-python>=1.4.0",
    "pyotp>=2.9.0",
    "yfinance>=0.2.0",
    "finnhub-python>=2.4.0",
    "uvicorn>=0.34.0",
    "starlette>=0.45.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "respx>=0.22.0",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
